Supply a client request id when calling into the SDK
The JS SDK doesn't send a client request id with the requests
(unlike the .NET SDK). Will follow up with the SDK team, but for now,
we will supply the client request id when calling into the SDK (via
custom headers in the request options).

Also ensure that setting values passed to the SDK are strings (SDK
calls will fail otherwise).